Login security update: sha1-challenge
Passwords are no longer sent plaintext, but rather a SHA1 hash
is calculated combined with a challenge, when the login form
contains a 'auth:challenge' field (default).

This requires passwords in db/auth.xml to be stored either in plaintext
(discouraged) or as a SHA1 hash.

NOTE that the sha1 hash of a password should be treated as a password!
This hash is however not transmitted over the network.

Updated the psp/auth.php login legacy code aswell: usernames
and passwords were prone to XPath code injection.