add psp_include_xml_eval for URI prefixes.
This XSLT accessible function will replace all {$foo}
references, such as:

	<a href="{$requestBaseURI}index.html">

into

	<a href="/index.html">

(where '/' is resolved by checking the global $request->requestBaseURI).

This is currently used to allow both 'native' PHP aswell as XSLT
to share files, such as a menu.xml.

The change is more or less equivalent with this code:

    ob_start();
    include_once( 'content/menu.xml' ); // doing it this way will execute any <?php code.
    $d = ob_get_contents();
    ob_end_clean();
    // finally evaluate any {$foo} expressions:
    global $request; extract( (array) $request );
    return eval("return <<<HTML\n$d\nHTML;\n");