lib/XML: XSLT caching update (XSLT disabled!)
re-using XSLTProcessor under certain circumstances
will simply close the HTTP connection.
Updated the caching to at least cache the DOMDocument of the XSL,
thus preserving at least the merged stylesheets (and the initial PSP XSL
application).
XSLTProcessor cache is currently disabled. When enabled however,
the registerFunctions and setParameters are called again to update
and/or register any new information (since a sheet that is applied
a second time later on will likely be for different content).