js/plot.js: refactor; better extreme calculation
This plot library allows to overlay several graphs onto one,
in which case all of them are normalized.
When showing individual graphs, it might sometimes be better
to normalize just the single graph for higher resolution.
This commit is a preparatory step in that direction by decoupling
the normalization factors and terms and applying them during
rendering, rather than modifying the original data.