add Session class with proper timeout; api/ async
The session_*  methods are now called through Session::*,
providing a hook to modify the $_SESSION contents on each
request, so that garbage collection will not close active sessions.

api/component AJAX methods open and close the session for the
duration of the need to access $_SESSION. This circumvents multiple
asynchronous AJAX calls from blocking on the session file.
Note that the default shutdown handler is assumed to close/release
the session file.